Work begun on an extended org to markdown converter
This would support fancier markdown engines. The idea was that
org-mode to Jekyll would go through markdown, rather than
HTML. However, it seems adequate and simpler to rely on org-mode's
HTML conversion engine and to simply export that for Jekyll.

Nevertheless, committing this work in case this was useful in the
future.